MoneyManagementApp/requirements.txt
flask
databricks-sql-connector
databricks-sdk